Sentiment Analysis of Hotel reviews
User comments/reviews together with scores "solicited" on hotels on the web are important sources of text based  information in "evaluating" hotel performance and planning travel for customers to remove uncertainty.  Travel planning and hotel booking on website has become an important commercial use of such user reviews.
In this social media age, sharing reviews on web has become a major tool in expressing customer thoughts about a particular product or Service. Such user reviews are relevant for purchase decisions since they are actual, based on experience and more detailed than reviews found in traditional printed hotel guides etc., and also not biased by marketing considerations as e.g. the hotels’ website pages or literature.
They also help hoteliers to boost ratings/revenue of hotels and serve as a way of engaging the users in the digital age. Therefore, keeping a tab on such comments, particularly negative, through "pick and choose" way and offering constructive response is important to the hotel management for quality control, improving reputation and communicating to the users that their opinions are valued. Needless to add that improved reputation translates to higher occupancy and better pricing. The hotel can increase its price appropriately based on sensitivity/elasticity of demand and still maintain the same occupancy or market share.
Performing analysis based on hotel reviews and segment of hotels (based on reviews and other features that you might feel appropriate to derive)such that the user might be informed about the similar hotels nearby.
Recommending hotels in a specific location to the users based on their previous preferences stated in user ratings.